change to bigint

+++ Models/DTO/AddToCart.cs
@@ -5,7 +5,7 @@
         public int userId { get; set; }
         public string name { get; set; }
         public string color { get; set; }
-        public int colorno { get; set; }
+        public long colorno { get; set; }
         public int quant { get; set; }
 
     }
