+++ Models/DTO/FiltersDTO.cs
@@ -6,6 +6,6 @@
         public decimal MaxPrice { get; set; }
         public int? Categories { get; set; }
         public string? Color { get; set; }
-        public int? ColorNo { get; set; }
+        public long? ColorNo { get; set; }
     }
 }
